Changes Skills to Outcomes, adds prerequisites
Changes the skills object to the outcomes array. Array makes
more sense as there are any number of possible outcomes to a
story and outcomes are not always based on just skills.

The new prerequisites array allows you to list conditions that
must be meet for that outcome to happen. Currently, I see this
as you need to meet any one of the prerequisites but we will
probably need to come up with a method to allow for needing to
meet multiple prerequisites.